usb: usbpsu: Fix MISRA C-2012 Rule 10.4 violation
Rule: MISRA C-2012 Rule 10.4
Essential type of the left hand operand (unsigned) is not
the same as that of the right operand "1U"(boolean).

This patch updates the variable with the required
type-cast.


Acked-for-series: Srinivas Goud <srinivas.goud@xilinx.com>

+++ XilinxProcessorIPLib/drivers/usbpsu/src/xusbpsu.c
@@ -175,7 +175,7 @@ s32 XUsbPsu_CfgInitialize(struct XUsbPsu *InstancePtr,
 	/*
 	 * Set connection speed based on EnableSuperSpeed parameter
 	 */
-	Speed = (ConfigPtr->EnableSuperSpeed == TRUE) ?
+	Speed = (ConfigPtr->EnableSuperSpeed == (u8)TRUE) ?
 			XUSBPSU_DCFG_SUPERSPEED : XUSBPSU_DCFG_HIGHSPEED;
 
 	/*
